# public/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Huiswerk App API Documentation",
        "description": "API documentation",
        "contact": {
            "email": "test@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Documentation page"
        }
    ],
    "paths": {
...
        "/wordlist/{id}": {
            "get": {
                "tags": [
                    "Wordlists"
                ],
                "summary": "Get a wordlist by ID",
                "description": "Returns a wordlist by ID",
                "operationId": "getWordlistById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of wordlist",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Wordlist"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Wordlist": {
                "title": "Wordlist",
                "description": "Wordlist Model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "string",
                        "example": "607d4567f533031b2627a1e3"
                    },
                    "words": {
                        "title": "Words",
                        "description": "List of words in a wordlist",
                        "type": "array",
                        "items": {
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "Aap"
                                },
                                "url": {
                                    "type": "string",
                                    "example": "*Base64 app image*"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Wordlist"
                }
            },
...
        },
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "description": "Key for the user",
                "name": "X-API-Key",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "apiKey": []
        }
    ]
}